style: 같은 폴더에서 관리하는 건 index로 공통 path 처리
